🐛 Fix writable_book_contents and written_book_contents item sub predicate format.
These predicates were introduced in 1.20.5-pre 1 and already in that version,
The format was as I documented it here.

The block post documented the format incorrectly as it was described by mcdoc
before.

I took this definition from the wiki and verified it working in 1.20.5-pre1,
but I wouldn't know if there are any other alternate versions, for example to
match fitlered text instead of raw.